Use float content bounds instead of int bounds

+++ src/gui/nlrsScrollPanel.h
@@ -10,7 +10,7 @@ class ScrollPanel : public Widget
 public:
     ScrollPanel(Widget* parent, IAllocator& allocator)
         : Widget(parent, allocator),
-        scrollBarWidth_(12),
+        scrollBarWidth_(12.f),
         feather_(4.f),
         radius_(3.f),
         scrollPosition_(0.0f)
@@ -21,7 +21,7 @@ public:
     void onRender() override;
     void onMouseScroll(i32 delta, Vec2i coordinates) override;
 
-    inline void setScrollBarWidth(i32 width)
+    inline void setScrollBarWidth(float width)
     {
         scrollBarWidth_ = width;
     }
@@ -39,7 +39,7 @@ public:
     }
 
 private:
-    i32 scrollBarWidth_;
+    float scrollBarWidth_;
     float feather_;
     float radius_;
     float scrollPosition_; // varies between 0, 1
